+++ packages/fiori/src/BarcodeScannerDialog.ts
@@ -2,14 +2,11 @@ import UI5Element from "@ui5/webcomponents-base/dist/UI5Element.js";
 import slot from "@ui5/webcomponents-base/dist/decorators/slot.js";
 import jsxRenderer from "@ui5/webcomponents-base/dist/renderer/JsxRenderer.js";
 import type I18nBundle from "@ui5/webcomponents-base/dist/i18nBundle.js";
-import Dialog from "@ui5/webcomponents/dist/Dialog.js";
-import Button from "@ui5/webcomponents/dist/Button.js";
-import BusyIndicator from "@ui5/webcomponents/dist/BusyIndicator.js";
+import type Dialog from "@ui5/webcomponents/dist/Dialog.js";
 import customElement from "@ui5/webcomponents-base/dist/decorators/customElement.js";
 import property from "@ui5/webcomponents-base/dist/decorators/property.js";
 import event from "@ui5/webcomponents-base/dist/decorators/event-strict.js";
 import i18n from "@ui5/webcomponents-base/dist/decorators/i18n.js";
-import { getI18nBundle } from "@ui5/webcomponents-base/dist/i18nBundle.js";
 import type { Result, Exception } from "@zxing/library";
 import type { Interval } from "@ui5/webcomponents-base/dist/types.js";
 // eslint-disable-next-line import/no-extraneous-dependencies
@@ -80,11 +77,6 @@ type BarcodeScannerDialogScanErrorEventDetail = {
 	renderer: jsxRenderer,
 	template: BarcodeScannerDialogTemplate,
 	styles: [BarcodeScannerDialogCss],
-	dependencies: [
-		Dialog,
-		BusyIndicator,
-		Button,
-	],
 })
 
 /**
@@ -190,10 +182,6 @@ class BarcodeScannerDialog extends UI5Element {
 		this._codeReader = new BrowserMultiFormatReader();
 		this._handleVideoPlayingBound = this._handleVideoPlaying.bind(this);
 		this._handleCaptureRegionBound = this._handleDrawCaptureRegion.bind(this);
-	}
-
-	static async onDefine() {
-		BarcodeScannerDialog.i18nBundle = await getI18nBundle("@ui5/webcomponents-fiori");
 	}
 
 	async onAfterRendering() {