+++ packages/fiori/src/ViewSettingsDialog.ts
@@ -8,16 +8,11 @@ import { isPhone } from "@ui5/webcomponents-base/dist/Device.js";
 import jsxRenderer from "@ui5/webcomponents-base/dist/renderer/JsxRenderer.js";
 import UI5Element from "@ui5/webcomponents-base/dist/UI5Element.js";
 import type { ChangeInfo } from "@ui5/webcomponents-base/dist/UI5Element.js";
-import Dialog from "@ui5/webcomponents/dist/Dialog.js";
-import Button from "@ui5/webcomponents/dist/Button.js";
-import Label from "@ui5/webcomponents/dist/Label.js";
-import ListItemGroup from "@ui5/webcomponents/dist/ListItemGroup.js";
-import List from "@ui5/webcomponents/dist/List.js";
+import type Dialog from "@ui5/webcomponents/dist/Dialog.js";
+import type List from "@ui5/webcomponents/dist/List.js";
 import type { ListItemClickEventDetail } from "@ui5/webcomponents/dist/List.js";
-import ListItemStandard from "@ui5/webcomponents/dist/ListItemStandard.js";
-import Title from "@ui5/webcomponents/dist/Title.js";
-import SegmentedButton from "@ui5/webcomponents/dist/SegmentedButton.js";
-import SegmentedButtonItem from "@ui5/webcomponents/dist/SegmentedButtonItem.js";
+import announce from "@ui5/webcomponents-base/dist/util/InvisibleMessage.js";
+import InvisibleMessageMode from "@ui5/webcomponents-base/dist/types/InvisibleMessageMode.js";
 
 import ViewSettingsDialogMode from "./types/ViewSettingsDialogMode.js";
 import "@ui5/webcomponents-icons/dist/sort.js";
@@ -36,6 +31,9 @@ import {
 	VSD_ORDER_ASCENDING,
 	VSD_ORDER_DESCENDING,
 	VSD_FILTER_BY,
+	VSD_SORT_TOOLTIP,
+	VSD_FILTER_TOOLTIP,
+	VSD_RESET_BUTTON_ACTION,
 } from "./generated/i18n/i18n-defaults.js";
 
 // Template
@@ -107,17 +105,6 @@ type VSDInternalSettings = {
 	renderer: jsxRenderer,
 	styles: viewSettingsDialogCSS,
 	template: ViewSettingsDialogTemplate,
-	dependencies: [
-		Button,
-		Title,
-		Dialog,
-		Label,
-		List,
-		ListItemStandard,
-		ListItemGroup,
-		SegmentedButton,
-		SegmentedButtonItem,
-	],
 })
 
 /**
@@ -356,6 +343,18 @@ class ViewSettingsDialog extends UI5Element {
 
 	get _sortByLabel() {
 		return ViewSettingsDialog.i18nBundle.getText(VSD_SORT_BY);
+	}
+
+	get _sortButtonTooltip() {
+		return ViewSettingsDialog.i18nBundle.getText(VSD_SORT_TOOLTIP);
+	}
+
+	get _filterButtonTooltip() {
+		return ViewSettingsDialog.i18nBundle.getText(VSD_FILTER_TOOLTIP);
+	}
+
+	get _resetButtonAction() {
+		return ViewSettingsDialog.i18nBundle.getText(VSD_RESET_BUTTON_ACTION);
 	}
 
 	get _isPhone() {
@@ -654,6 +653,7 @@ class ViewSettingsDialog extends UI5Element {
 		this._restoreSettings(this._initialSettings);
 		this._recentlyFocused = this._sortOrder!;
 		this._focusRecentlyUsedControl();
+		announce(this._resetButtonAction, InvisibleMessageMode.Polite);
 	}
 
 	/**