+++ packages/main/src/Toolbar.ts
@@ -34,11 +34,10 @@ import type ToolbarSeparator from "./ToolbarSeparator.js";
 import {
 	getRegisteredToolbarItem,
 	getRegisteredStyles,
-	getRegisteredDependencies,
 } from "./ToolbarRegistry.js";
 
-import Button from "./Button.js";
-import Popover from "./Popover.js";
+import type Button from "./Button.js";
+import type Popover from "./Popover.js";
 
 type ToolbarMinWidthChangeEventDetail = {
 	minWidth: number,
@@ -174,15 +173,6 @@ class Toolbar extends UI5Element {
 			ToolbarCss,
 			ToolbarPopoverCss,
 			...styles,
-		];
-	}
-
-	static get dependencies() {
-		const deps = getRegisteredDependencies();
-		return [
-			Popover,
-			Button,
-			...deps,
 		];
 	}
 
@@ -363,6 +353,9 @@ class Toolbar extends UI5Element {
 	 */
 
 	processOverflowLayout() {
+		if (this.offsetWidth === 0) {
+			return;
+		}
 		const containerWidth = this.offsetWidth - this.padding;
 		const contentWidth = this.itemsWidth;
 		let overflowSpace = contentWidth - containerWidth + this.overflowButtonSize;